Extract rust-embed-rocket crate
The code in there is shared between both binaries. If we want to
extract a dedicated crate per binary, this code needs to go into
a dedicated crate.
