+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["daemon", "rocket-basicauth"]
+members = ["daemon", "rocket-basicauth", "rust-embed-rocket"]
 resolver = "2"
 
 [patch.crates-io]
