+++ daemon/Cargo.toml
@@ -28,6 +28,7 @@ reqwest = { version = "0.11", default-features = false, features = ["json", "rus
 rocket = { version = "0.5.0-rc.1", features = ["json", "uuid"] }
 rocket-basicauth = { path = "../rocket-basicauth" }
 rust-embed = "6.3"
+rust-embed-rocket = { path = "../rust-embed-rocket" }
 rust_decimal = "1.21"
 rust_decimal_macros = "1.20"
 semver = { version = "1.0.4", features = ["serde"] }
