+++ daemon/src/routes_maker.rs
@@ -12,7 +12,6 @@ use daemon::oracle;
 use daemon::projection::Cfd;
 use daemon::projection::CfdAction;
 use daemon::projection::Feeds;
-use daemon::routes::EmbeddedFileExt;
 use daemon::to_sse_event::ToSseEvent;
 use daemon::wallet;
 use daemon::MakerActorSystem;
@@ -26,6 +25,7 @@ use rocket::serde::json::Json;
 use rocket::State;
 use rocket_basicauth::Authenticated;
 use rust_embed::RustEmbed;
+use rust_embed_rocket::EmbeddedFileExt;
 use serde::Deserialize;
 use std::borrow::Cow;
 use std::path::PathBuf;
