+++ daemon/src/routes_taker.rs
@@ -11,7 +11,6 @@ use daemon::oracle;
 use daemon::projection;
 use daemon::projection::CfdAction;
 use daemon::projection::Feeds;
-use daemon::routes::EmbeddedFileExt;
 use daemon::to_sse_event::Heartbeat;
 use daemon::to_sse_event::ToSseEvent;
 use daemon::wallet;
@@ -26,6 +25,7 @@ use rocket::serde::json::Json;
 use rocket::State;
 use rocket_basicauth::Authenticated;
 use rust_embed::RustEmbed;
+use rust_embed_rocket::EmbeddedFileExt;
 use serde::Deserialize;
 use serde::Serialize;
 use std::borrow::Cow;
